Add more samples in main

+++ thread-pool/build.sh
@@ -2,4 +2,6 @@
 
 g++ thread_test.cpp -Iinclude -lpthread -std=c++11 -o thread_test
 
-mkdir -p bin && mv ./thread_test ./bin
+if [ -e thread_test ]; then
+    mkdir -p bin && mv ./thread_test ./bin
+fi
